+++ metadata.rb
@@ -3,7 +3,7 @@ maintainer 'Mariani Lucas'
 maintainer_email 'marianilucas@gmail.com'
 license 'Apache 2.0'
 description 'Installs/Configures yum-remi'
-version '0.1.0'
+version '1.0.0'
 
 supports 'centos'
 supports 'redhat'