+++ recipes/remi-test.rb
@@ -1,8 +1,8 @@
 #
 # Author:: Sean OMeara (<sean@chef.io>)
-# Recipe:: yum-remi::remi
+# Recipe:: yum-remi::remi-test
 #
-# Copyright 2015, Chef
+# Copyright 2016, Chef
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-include_recipe 'yum-epel' unless node['platform'] == 'fedora'
+include_recipe 'yum-remi::remi'
 
 %w(remi-test remi-test-debuginfo).each do |repo|
   next unless node['yum'][repo]['managed']