+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 yum-remi Cookbook CHANGELOG
 ===========================
 
+v1.0.0
+------
+Removed support for Fedora linux older than 21.
+Updated for most recent remi repositories.
+Added remi-php70.
+
 v0.1.0
 ------
 initial release